feat: add suppress_noisy_loggers function to manage verbose HTTP loggers
