+++ src/sparkrun/bootstrap.py
@@ -15,21 +15,6 @@ logger = logging.getLogger(__name__)
 
 EXT_RUNTIME = "sparkrun.runtime"
 
-# Loggers that produce excessive output at DEBUG/INFO level.
-_NOISY_LOGGERS = (
-    "httpx",
-    "httpcore.http11",
-    "httpcore.connection",
-    "urllib3.connectionpool",
-)
-
-
-def suppress_noisy_loggers() -> None:
-    """Suppress verbose HTTP/transport loggers."""
-    for name in _NOISY_LOGGERS:
-        logging.getLogger(name).setLevel(logging.WARNING)
-
-
 # Module-level singleton for the sparkrun Variables instance
 _variables: Variables | None = None
 
@@ -58,6 +43,7 @@ def init_sparkrun(v: Variables | None = None, log_level: str = "WARNING") -> Var
         v = init_framework_desktop("sparkrun", log_level=log_level, fault_handler=False, shutdown_hooks=False, fixed_logger=logger)
 
         # suppress noisy loggers (separate from our logging level)
+        from sparkrun.utils import suppress_noisy_loggers
         suppress_noisy_loggers()
 
     _variables = v
